:recycle: refactor: refactiring requestManager and tests
