+++ tests/033-EditeIssueName.test.js
@@ -1,11 +1,13 @@
-require('dotenv').config();
-const requestManager = require('./../utils/requestManager');
+const env = require('../core/configs/environments');
+
+const requestManager = require('../core/utils/requestManager');
+let createdIssueId = '';
+
 const basicAuth =
   'Basic ' +
   Buffer.from(
-    `${process.env.JIRA_EMAIL}:${process.env.JIRA_API_TOKEN}`,
+    `${env.environment.username}:${env.environment.api_token}`,
   ).toString('base64');
-let createdIssueId;
 
 beforeEach(async () => {
   try {
