+++ tests/039-EditeCommentIssue.test.js
@@ -1,10 +1,13 @@
-require('dotenv').config();
-const requestManager = require('./../utils/requestManager');
+const env = require('../core/configs/environments');
+
+const requestManager = require('../core/utils/requestManager');
+
 const basicAuth =
   'Basic ' +
   Buffer.from(
-    `${process.env.JIRA_EMAIL}:${process.env.JIRA_API_TOKEN}`,
+    `${env.environment.username}:${env.environment.api_token}`,
   ).toString('base64');
+
 let createdIssueId;
 
 beforeEach(async () => {
